Import SMILemu, add everything needed to make it build.
Couldn't get the original packages to build cleanly,
so this is a workaround to be able to build this from
source. For me there is no need to separate things into
multiple packages, nor keep the Android build around.

An fixes can be backported to the canonical repository
if/when needed.