Add correct typeface for typewriter.
An IBM Model B typewriter was used, and judging by printouts
the typeface used was Bookface Academic. Later IBM typewriters
such as the Selectric used this typeface. Here a subset of
Selectric Bookface 72 has been retraced from old printouts.
I call this subset SMIL so as not to confuse it with the full
typeface.